refactor: make Framework folder structure more intuitive
