+++ app/Framework/Models/Model.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Framework\Classes;
+namespace App\Framework\Models;
 
 abstract class Model
 {
