+++ app/Framework/View/ErrorPage.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Framework\View;
+
+class ErrorPage
+{
+
+}
