+++ resources/views/anao/list.php
@@ -1,0 +1,1 @@
+<?php
