+++ resources/views/auth/login.php
@@ -1,0 +1,1 @@
+<?php
